+++ app/src/main/java/budget/application/model/dto/AccountRequest.java
@@ -1,4 +1,6 @@
 package budget.application.model.dto;
 
+import java.math.BigDecimal;
+
 public record AccountRequest(
-    String name, String accountType, String bankName, double openingBalance, String status) {}
+    String name, String accountType, String bankName, BigDecimal openingBalance, String status) {}
