+++ app/src/main/java/budget/application/model/dto/AccountResponse.java
@@ -1,6 +1,7 @@
 package budget.application.model.dto;
 
 import io.github.bibekaryal86.shdsvc.dtos.ResponseMetadata;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -10,6 +11,6 @@ public record AccountResponse(List<Account> data, ResponseMetadata metadata) {
       String name,
       String accountType,
       String bankName,
-      double openingBalance,
+      BigDecimal openingBalance,
       String status) {}
 }
