+++ app/src/main/java/budget/application/model/dto/TransactionItemRequest.java
@@ -1,6 +1,7 @@
 package budget.application.model.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record TransactionItemRequest(
-    UUID transactionId, UUID categoryId, String label, double amount, String expType) {}
+    UUID transactionId, UUID categoryId, String label, BigDecimal amount, String expType) {}
